refactor: getToken to utils directory
